Deduce types

+++ tests/irix-g/output/int-float.c
@@ -1,7 +1,7 @@
 test(f32 f12, s32 a1, f32 arg8) {
-    (s32) D_410130 = (s32) f12;
-    (f32) D_410134 = (f32) a1;
-    (f32) arg8 = (arg8 + 5.0f);
+    D_410130 = (s32) f12;
+    D_410134 = (f32) a1;
+    arg8 = (f32) (arg8 + 5.0f);
     return;
     // (possible return value: ((f32) (u32) (a3 + 3) + arg8))
 }
