+++ tests/irix-o2/output/test.c
@@ -1,5 +1,5 @@
 test(void) {
-    (s32) D_4100E0 = 4;
+    D_4100E0 = 4;
     func_004000B0();
     // (possible return value: func_004000B0())
 }
